+++ src/lib/libast/cdt/dttree.c
@@ -14,6 +14,7 @@
 *                  David Korn <dgk@research.att.com>                   *
 *                   Phong Vo <kpv@research.att.com>                    *
 *                  Martijn Dekker <martijn@inlv.org>                   *
+*            Johnothan King <johnothanking@protonmail.com>             *
 *                                                                      *
 ***********************************************************************/
 #include	"dthdr.h"
@@ -207,7 +208,7 @@ static void* tstat(Dt_t* dt, Dtstat_t* st)
 		/**/DEBUG_ASSERT((dt->data->type&DT_SHARE) || size == dt->data->size);
 		st->meth  = dt->meth->type;
 		st->size  = size;
-		st->space = sizeof(Dttree_t) + (dt->disc->link >= 0 ? 0 : size*sizeof(Dthold_t));
+		st->space = (ssize_t)(sizeof(Dttree_t) + (dt->disc->link >= 0 ? 0 : (size_t)size*sizeof(Dthold_t)));
 		return (void*)size;
 	}
 }
