+++ src/lib/libast/include/glob.h
@@ -2,7 +2,7 @@
 *                                                                      *
 *               This software is part of the ast package               *
 *          Copyright (c) 1985-2011 AT&T Intellectual Property          *
-*          Copyright (c) 2020-2022 Contributors to ksh 93u+m           *
+*          Copyright (c) 2020-2025 Contributors to ksh 93u+m           *
 *                      and is licensed under the                       *
 *                 Eclipse Public License, Version 2.0                  *
 *                                                                      *
@@ -81,7 +81,7 @@ struct _glob_
 	char*		(*gl_nextdir)(glob_t*, char*);
 	int		gl_status;
 	unsigned long	gl_version;
-	unsigned short	gl_extra;
+	size_t		gl_extra;
 
 #ifdef _GLOB_PRIVATE_
 	_GLOB_PRIVATE_
